Moved js files into main dir
