+++ index.js
@@ -1,7 +1,7 @@
 const NewsModel = require('./newsModel');
 const NewsView = require('./newsView');
-const NewsApi = require('./newsApi')
+const NewsApi = require('./newsApi');
 
 const model = new NewsModel;
-const api = new NewsAPI;
+const api = new NewsApi;
 const view = new NewsView(model, api);
