+++ newsModel.js
