+++ newsModel.test.js
