+++ newsView.js
